pyproject.toml
[tool.poetry]
name = "translations_service"
version = "0.1.0"
description = ""
authors = ["IvanGlebov <glebov.vanya8@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"
pandas = "^1.4.3"
gspread = "^5.4.0"
google = "^3.0.0"
google-auth-oauthlib = "^0.5.2"
google-api-python-client = "^2.56.0"
GitPython = "^3.1.27"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
